Refactor `src/core/property.rs` God Module into a cohesive module directory
